% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{algorithm-class}
\alias{algorithm-class}
\title{Virtual Parent Class Algorithm}
\description{
It will use when other algorithms are implemented. Equivalent to a Abstract
class in other languages.
}
