% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_crowding_distance.R
\name{calc_norm_pref_distance}
\alias{calc_norm_pref_distance}
\title{Calculate Normalized Preference Distance
Computes the weighted normalized Euclidean distance between a set of fitness
vectors and a set of reference points.}
\usage{
calc_norm_pref_distance(fitness, ref_points, weight, ideal_point, nadir_point)
}
\arguments{
\item{fitness}{A matrix of fitness values.}

\item{ref_points}{A matrix of reference points.}

\item{weight}{A numeric vector of weights for each objective.}

\item{ideal_point}{A numeric vector of ideal point values.}

\item{nadir_point}{A numeric vector of nadir point values.}
}
\value{
A matrix of distances where element (i, j) is the distance from
fitness to ref_points.
}
\description{
Calculate Normalized Preference Distance
Computes the weighted normalized Euclidean distance between a set of fitness
vectors and a set of reference points.
}
