% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowding_distance.R
\name{crowding_distance}
\alias{crowding_distance}
\title{Calculation of Crowding Distance}
\usage{
crowding_distance(object, nObj)
}
\arguments{
\item{object, nObj}{An object of class 'nsga2', usually resulting from a call to function nsga2.
Fitness Function Objective Numbers}
}
\value{
A vector with the crowding-distance between individuals of a population.
}
\description{
A Crowded-comparison approach.
}
\details{
The crowded-comparison operator guides the selection process at the various
stages of the algorithm toward a uniformly spread-out Pareto-optimal front
}
\references{
K. Deb, A. Pratap, S. Agarwal and T. Meyarivan, 'A fast and elitist multiobjective
genetic algorithm: NSGA-II,' in IEEE Transactions on Evolutionary Computation,
vol. 6, no. 2, pp. 182-197, April 2002, doi: 10.1109/4235.996017.
}
\seealso{
\code{\link[=non_dominated_fronts]{non_dominated_fronts()}}
}
\author{
Francisco Benitez
}
