% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nsga2.R
\docType{methods}
\name{getCrowdingDistance}
\alias{getCrowdingDistance}
\alias{getCrowdingDistance,nsga2-method}
\title{Accessor methods to the crowding distance for NSGA-II results}
\usage{
getCrowdingDistance(obj)

\S4method{getCrowdingDistance}{nsga2}(obj)
}
\arguments{
\item{obj}{an object resulting from the execution of NSGA-II algorithm}
}
\value{
Returns a vector with the crowding distances of class nsga2. See \linkS4class{nsga2}
for a description of available slots information.
}
\description{
Accessor methods to the crowding distance for NSGA-II results
}
\examples{
# Where 'out' is an object resulting from the execution of the NSGA-II algorithm.
#
# getCrowdingDistance(out)
#

}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
