% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getDummyFitness-method}
\alias{getDummyFitness-method}
\title{Accessor methods to the dummy fitness for NSGA-I results}
\usage{
getDummyFitness(obj)
}
\arguments{
\item{obj}{an object resulting from the execution of NSGA-I algorithm}
}
\value{
Returns a matrix with the dummy fitness of class nsga1. See \linkS4class{nsga1}
for a description of available slots information.
}
\description{
Accessor methods to the dummy fitness for NSGA-I results
}
\examples{
# Where 'out' is an object resulting from the execution of the NSGA-I algorithm.
#
# getDummyFitness(out)
#

}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
