% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getMetrics}
\alias{getMetrics}
\alias{getMetrics,nsga-method}
\alias{getMetrics,nsga,nsga-method}
\title{Accessor methods to the metrics evaluated during execution}
\usage{
getMetrics(obj)

\S4method{getMetrics}{nsga}(obj)
}
\arguments{
\item{obj}{an object resulting from the execution of NSGA-I, NSGA-II or NSGA-III
algorithm. During the execution of the performance metrics must be evaluated.}
}
\value{
A dataframe with performance metrics evaluated iteration by iteration.
}
\description{
Accessor methods to the metrics evaluated during execution
}
\examples{
# Where 'out' is an object resulting from the execution of the rmoo.
#
# metrics_result <- getMetrics(out)
#
# metrics_result

}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
