% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getPopulation-method}
\alias{getPopulation-method}
\title{Accessor methods to the population for rmoo results}
\usage{
getPopulation(obj)
}
\arguments{
\item{obj}{an object resulting from the execution of NSGA-I, NSGA-II or NSGA-III
algorithm}
}
\value{
Prints the resulting population and when the result of the method-call
is assigned to a variable, the population is stored as a data frame.
See \linkS4class{nsga1} \linkS4class{nsga2}, \linkS4class{nsga3} for a description of available
slots information.
}
\description{
Accessor methods to the population for rmoo results
}
\examples{
# Where 'out' is an object resulting from the execution of rmoo.
#
# population_result <- getPopulation(out)
#
# population_result

}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
