% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kroA100}
\alias{kroA100}
\title{KROA100}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{COORD}{City Coordinates}
\item{SECTION}{City Section}
}
}
\usage{
kroA100
}
\description{
A dataset containing the coord and section of 100 cities
}
\references{
Reinelt, G. (1991). TSPLIB—A traveling salesman problem library.
ORSA journal on computing, 3(4), 376-384
}
\keyword{datasets}
