% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{nsga1-class}
\alias{nsga1-class}
\title{Class 'nsga1'}
\description{
The class 'nsga1' is instantiated within the execution of rmoo and will be
returned as a result of it. All data generated during execution will be
stored in it.
}
\section{Slots}{

\describe{
\item{\code{dumFitness}}{a large dummy fitness value assigned to individuals from
the nondominated front.}

\item{\code{dShare}}{the maximun phenotypic distance allowed between any two
individuals to become members of a niche.}

\item{\code{deltaDummy}}{value to decrease the dummy fitness of individuals by
non-dominated fronts.}
}}

\examples{
showClass('nsga1')
}
