% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{nsga2-class}
\alias{nsga2-class}
\title{Class 'nsga2'}
\description{
The class 'nsga2' is instantiated within the execution of rmoo and will be
returned as a result of it. All data generated during execution will be
stored in it.
}
\section{Slots}{

\describe{
\item{\code{crowdingDistance}}{Crowding-comparison approach to estimate of the
perimeter of the cuboid formed by using the nearest neighbors as the vertices.}
}}

\examples{
showClass('nsga2')
}
