% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{nsga3-class}
\alias{nsga3-class}
\title{Class 'nsga3'}
\description{
The class 'nsga3' is instantiated within the execution of rmoo and will be
returned as a result of it. All data generated during execution will be
stored in it.
}
\section{Slots}{

\describe{
\item{\code{ideal_point}}{Nadir point estimate used as lower bound in normalization.}

\item{\code{worst_point}}{Worst point generated over generations.}

\item{\code{smin}}{Index used to obtain the extreme points.}

\item{\code{extreme_points}}{are selected using the ASF in the (\code{\link[=PerformScalarizing]{PerformScalarizing()}}).
Necessary in the  nadir point generation.}

\item{\code{worst_of_population}}{The worst individuals generated by objectives in
the current generation.}

\item{\code{worst_of_front}}{The worst individuals in the first front generated by
objectives in the current generation.}

\item{\code{nadir_point}}{Nadir point estimate used as upper bound in normalization.}

\item{\code{reference_points}}{NSGA-III uses a predefined set of reference points to
ensure diversity in obtained solutions.
The chosen refenrece points can be predefined in structured manner or
supplied by the user. We use the Das and Dennis procedure.}
}}

\examples{
showClass('nsga3')
}
