% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga3r.R
\docType{package}
\name{nsga3r-package}
\alias{nsga3r}
\alias{nsga3r-package}
\title{nsga3r: Family of Non-Dominated Genetic Algorithms}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A multiobjective optimization package based on K. Deb's 
    algorithm and inspired by Luca Scrucca's GA package <10.32614/RJ-2017-008>. 
    The nsga3r package is a framework for multi- and many-objective optimization, 
    allowing to work with representation of real numbers, permutations and binaries, 
    offering a high range of configurations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benitezfj/nsga3r}
  \item Report bugs at \url{https://github.com/benitezfj/nsga3r/issues}
}

}
\author{
\strong{Maintainer}: Francisco Benitez \email{benitez.fj@hotmail.com}

Authors:
\itemize{
  \item Diego Pinto Roa \email{dpinto@pol.una.py} (\href{https://orcid.org/0000-0003-2479-9876}{ORCID})
}

}
\keyword{internal}
