% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{print}
\alias{print}
\title{Methods for Function 'print' in Package 'rmoo'.}
\arguments{
\item{x, y}{Objects of either class \linkS4class{nsga1},
\linkS4class{nsga2},  or \linkS4class{nsga3}.}

\item{...}{other arguments passed on to methods}
}
\value{
Print the slots and relevant values of the object.
}
\description{
Method used to print the slots and relevant values of the object.
}
\examples{
# Where 'out' is an object of class nsga1, nsga2, or nsga3
#
# print(out)

}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
