% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{progress-method}
\alias{progress-method}
\title{Methods for Function 'progress' in Package 'rmoo'}
\usage{
progress(object, ...)
}
\arguments{
\item{...}{other arguments passed on to methods. Passing \code{"reference_dirs"}
as arguments will evaluate the performance metrics Hypervolumen,
Generational Distance, and Inverse Generational Distance.}

\item{x, y}{Objects of either class \linkS4class{nsga1},
\linkS4class{nsga2},  or \linkS4class{nsga3}.}
}
\value{
A list of length equal to the number of iterations, where the progress made during execution is saved.
}
\description{
Method used to save the progress of the evaluation results, similar to the
summary method. Passing additional arguments to the progress method evaluates
performance metrics per iteration. This method cannot be called outside of
rmoo execution.
}
\examples{
# Where 'out' is an object of class nsga1, nsga2, or nsga3, and callArgs are
# the additional arguments passed when calling the rmoo function, for the
# evaluation of performance metrics, reference points are expected to be passed
# as an argument to reference_dirs.
#
# progress(object, callArgs)
#

}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
