% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_point_multi_layer.R
\name{reference_point_multi_layer}
\alias{reference_point_multi_layer}
\title{Determination of Multi-layer Reference Points}
\usage{
reference_point_multi_layer(...)
}
\arguments{
\item{...}{The different layers provided by the user}
}
\value{
A matrix with the multi-layer reference points
}
\description{
A implementation of Multi-layer Reference Points Generation.
}
\details{
The Multi-layer reference point implementation is based on Blank and Deb's
pymoo library, the approach generates different layers of references point
at different scales, provided by the user.
}
\references{
J. Blank and K. Deb, "Pymoo: Multi-Objective Optimization in
Python," in IEEE Access, vol. 8, pp. 89497-89509, 2020,
doi: 10.1109/ACCESS.2020.2990567.

Das, Indraneel & Dennis, J. (2000). Normal-Boundary Intersection: A New
Method for Generating the Pareto Surface in Nonlinear Multicriteria
Optimization Problems. SIAM Journal on Optimization. 8.
10.1137/S1052623496307510.
}
\seealso{
\code{\link[=generate_reference_points]{generate_reference_points()}} and \code{\link[=get_fixed_rowsum_integer_matrix]{get_fixed_rowsum_integer_matrix()}}
}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
