% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticoperator.R
\name{rmoo_tourSelection}
\alias{rmoo_tourSelection}
\alias{rmooreal_tourSelection}
\alias{rmoobin_tourSelection}
\alias{rmooperm_tourSelection}
\title{Tournament Selection}
\usage{
rmoo_tourSelection(object, k = 2, ...)

rmooreal_tourSelection(object, k = 2, ...)

rmoobin_tourSelection(object, k = 2, ...)

rmooperm_tourSelection(object, k = 2, ...)
}
\arguments{
\item{object}{MOEA object with slots population, fitness, front, popSize.}

\item{k}{Tournament size.}

\item{...}{Argument which all the values necessary for the configuration
will be passed as parameters. The user is encouraged to see the documentations.}
}
\value{
List with \code{population} and \code{fitness} subsets.
}
\description{
Binarily o por pref
}
\examples{
\dontrun{
# Creamos un "dummy" objeto mínimo
object <- list(
  population       = matrix(runif(20), nrow=5),
  fitness          = matrix(runif(10), nrow=5),
  front            = matrix(sample(1:2, 5, TRUE), ncol=1),
  crowdingDistance = runif(5),
  popSize          = 5
)
class(object) <- "nsga2"
# Llamamos al selector
sel <- rmoo_tourSelection(object, k = 2)
str(sel)
}
}
