% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing.R
\name{sharing}
\alias{sharing}
\title{Calculation of Dummy Fitness}
\usage{
sharing(object)
}
\arguments{
\item{object}{An object of class 'nsga', usually resulting from a call to
function nsga. Fitness Function Objective Numbers.}
}
\value{
A vector with the dummy fitness.
}
\description{
Calculate of sharing distance and dummy fitness
}
\details{
The sharing distance operator guides the selection process at the various
stages of the algorithm toward a uniformly spread-out Pareto-optimal front
}
\references{
N. Srinivas and K. Deb, 'Multiobjective Optimization Using
Nondominated Sorting in Genetic Algorithms,' in Evolutionary Computation,
vol. 2, no. 3, pp. 221-248, Sept. 1994, doi: 10.1162/evco.1994.2.3.221.
}
\seealso{
\code{\link[=non_dominated_fronts]{non_dominated_fronts()}}
}
\author{
Francisco Benitez
}
