% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{startParallel}
\alias{startParallel}
\title{Start Parallel Backend for rmoo Package}
\usage{
startParallel(parallel = TRUE, ...)
}
\arguments{
\item{parallel}{Logical, numeric, character, or a cluster object. If \code{TRUE}, uses all detected cores.
If numeric, it specifies the number of cores. If character, should be \code{"snow"} or \code{"multicore"}.
If a cluster object, it will use that cluster.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class \code{logical} with attributes:
\itemize{
\item \code{type}: cluster type ("snow" or "multicore")
\item \code{cores}: number of cores used
\item \code{cluster}: the cluster object created or passed
}
}
\description{
This function sets up parallel computing using the \code{parallel} and \code{doParallel} packages.
It supports both "snow" (PSOCK) and "multicore" backends depending on the OS.
}
\examples{
\dontrun{
  cl <- startParallel(TRUE)
  stopParallel(attr(cl, "cluster"))
}

}
