% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_crowding_distance.R
\name{modifiedCrowdingDistance}
\alias{modifiedCrowdingDistance}
\title{Calculation of Modified Crowding Distance}
\usage{
modifiedCrowdingDistance(
  object,
  epsilon = 0.001,
  weights = NULL,
  normalization = "front",
  extreme_points_as_ref_dirs = FALSE
)
}
\arguments{
\item{object, nObj}{An object of class 'nsga2', usually resulting from a call
to function nsga2. Fitness Function Objective Numbers}
}
\value{
A vector with the crowding-distance between individuals of a population.
}
\description{
A Crowded-comparison approach.
}
\details{
The crowded-comparison operator guides the selection process at the various
stages of the algorithm toward a uniformly spread-out Pareto-optimal front
}
\references{
Kalyanmoy Deb and J. Sundar. 2006. Reference point based
multi-objective optimization using evolutionary algorithms. In Proceedings of
the 8th annual conference on Genetic and evolutionary computation (GECCO '06).
Association for Computing Machinery, New York, NY, USA, 635–642.
doi: 10.1145/1143997.1144112
}
\seealso{
\code{\link[=rnsga2]{rnsga2()}}
}
\author{
Francisco Benitez
\email{benitezfj94@gmail.com}
}
