% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_Imports.R
\docType{package}
\name{nsga3r-package}
\alias{nsga3r}
\alias{nsga3r-package}
\title{nsga3r: Non-Dominated Genetic Algorithms - III}
\description{
A multi-objective optimization package based on the
    algorithm of K. Deb and H. Jain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benitezfj/nsga3r}
  \item Report bugs at \url{https://github.com/benitezfj/nsga3r/issues}
}

}
\author{
\strong{Maintainer}: Francisco Benitez \email{benitez.fj@hotmail.com}

Authors:
\itemize{
  \item Diego Pinto Roa \email{dpinto@pol.una.py} (\href{https://orcid.org/0000-0003-2479-9876}{ORCID})
}

}
\keyword{internal}
