% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberOrNAOrMatrix-class.R
\docType{class}
\name{numberOrNAOrMatrix-class}
\alias{numberOrNAOrMatrix-class}
\title{Virtual Class 'numberOrNAOrMatrix - Simple Class for subassigment Values'}
\description{
The class 'numberOrNAOrMatrix' is a simple class union (\code{\link[=setClassUnion]{setClassUnion()}}) of 'numeric', 'logical' and 'matrix'.
}
\section{Objects from the Class}{

Since it is a virtual Class, no objects may be created from it.
}

\examples{
showClass('numberOrNAOrMatrix')
}
