\name{performance_metrics}
\alias{performance_metrics}
%
\alias{generational_distance}

\title{Objective Values performance metrics}

\description{Functions to evaluate the quality of the results obtained by the algorithms, evaluating their diversity and convergence, providing or not some parameters to compare.}

\usage{
  generational_distance(fitness, reference_points)
}

\arguments{
  \item{fitness}{Objective values generated by the algorithm.}
  
  \item{reference_points}{Optimal points to achieve.}
}

\value{A vector with the measurement parameter.}

\references{
  Lamont, G., & Veldhuizen, D.V. (1999). Multiobjective evolutionary algorithms: classifications, analyses, and new innovations.

}

\author{Francisco Benitez}
