% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoo.R
\docType{package}
\name{rmoo-package}
\alias{rmoo-package}
\alias{_PACKAGE}
\title{rmoo: Multi-Objective Optimization in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'rmoo' package is a framework for multi- and many-objective optimization, which allows researchers and users versatility in parameter configuration, as well as tools for analysis, replication and visualization of results. The 'rmoo' package was built as a fork of the 'GA' package by Luca Scrucca(2017) \doi{10.32614/RJ-2017-008} and implementing the Non-Dominated Sorting Genetic Algorithms proposed by K. Deb's.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Evolutionary-Optimization-Laboratory/rmoo/}
  \item Report bugs at \url{https://github.com/Evolutionary-Optimization-Laboratory/rmoo/issues/}
}

}
\author{
\strong{Maintainer}: Francisco Benitez \email{benitezfj94@gmail.com}

Authors:
\itemize{
  \item Diego P. Pinto-Roa \email{dpinto@pol.una.py} (\href{https://orcid.org/0000-0003-2479-9876}{ORCID})
}

}
\keyword{internal}
