% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoo.R
\docType{package}
\name{rmoo-package}
\alias{rmoo}
\alias{rmoo-package}
\title{rmoo: 'Non-Dominated Genetic Algorithms' Family}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A multiobjective optimization package based on K. Deb's 
    algorithm and inspired in 'GA' package by Luca Scrucca (2017) <DOI:10.32614/RJ-2017-008>. 
    The 'rmoo' package is a framework for multi- and many-objective optimization, 
    allowing to work with representation of real numbers, permutations and 
    binaries, offering a high range of configurations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benitezfj/rmoo}
  \item Report bugs at \url{https://github.com/benitezfj/rmoo/issues}
}

}
\author{
\strong{Maintainer}: Francisco Benitez \email{benitezfj94@gmail.com}

Authors:
\itemize{
  \item Diego Pinto Roa \email{dpinto@pol.una.py} (\href{https://orcid.org/0000-0003-2479-9876}{ORCID})
}

}
\keyword{internal}
