% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{rnsga2-class}
\alias{rnsga2-class}
\title{Class 'rnsga2'}
\description{
The class 'rnsga2' is instantiated within the execution of rmoo and will be
returned as a result of it. All data generated during execution will be
stored in it.
}
\section{Slots}{

\describe{
\item{\code{crowdingDistance}}{Crowding-comparison approach to estimate of the
perimeter of the cuboid formed by using the nearest neighbors as the vertices.}

\item{\code{reference_points}}{R-NSGA-II uses a set of reference points defined by the user to
ensure diversity in obtained solutions.}

\item{\code{extreme_points}}{are selected using the ASF in the (\code{\link[=PerformScalarizing]{PerformScalarizing()}}).
Necessary in the  nadir point generation.}

\item{\code{smin}}{Index used to obtain the extreme points.}
}}

\examples{
showClass('rnsga2')
}
